refactor: AdminController 생성 truncate.sql 수정
