+++ src/test/java/wooteco/subway/admin/repository/StationRepositoryTest.java
@@ -3,13 +3,11 @@ package wooteco.subway.admin.repository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.data.jdbc.DataJdbcTest;
-import org.springframework.test.context.jdbc.Sql;
 import wooteco.subway.admin.domain.Station;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DataJdbcTest
-@Sql("/truncate.sql")
 public class StationRepositoryTest {
     @Autowired
     private StationRepository stationRepository;
