+++ src/test/resources/truncate.sql
@@ -1,2 +1,2 @@
-DELETE FROM STATION;
-DELETE FROM LINE;
+TRUNCATE TABLE LINE;
+TRUNCATE TABLE STATION;
